3rd initial development commit

+++ UndiscordOCR.py
@@ -14,8 +14,12 @@ config = ConfigParser()
 config.read("config.ini")
 config_data = config['default']
 token = config_data['token']
+channel_id = config_data['channel_id']
+history_deletion = config_data['history_deletion']
+store_flagged = config_data['store_flagged_messeges']
 
-word = ['BANSOLA', 'Bansola', 'bansola','8ur3j85hy98']
+word = config_data['word']
+word.replace(",","")
 
 def words_in_string(word_list, text):
     return set(word_list).intersection(text.split())
@@ -23,7 +27,7 @@ def words_in_string(word_list, text):
 
 
 async def delete_history():
-    channel = client.get_channel(849794290317525014)
+    channel = client.get_channel(channel_id)
     messages = [message async for message in channel.history(limit=100000)]
     
     i =0
@@ -64,14 +68,18 @@ async def delete_history():
 #WATCH FOR MESSAGES IN REAL TIME
 class MyClient(discord.Client):
     async def on_ready(self):
+        await MyClient.change_presence(status=discord.Status.invisible)
         print(f'Logged on as {self.user}!')
-        await delete_history()
+
+
+        if history_deletion == 'True':
+            await delete_history()
 
     async def on_message(self, message):
         #if message.author.id == (1203429051226787890):
                      #await message.delete()   
 
-        print(f'Message from {message.author}: {message.content}')
+        #print(f'Message from {message.author}: {message.content}')
         #with open('log.txt', 'a') as f:
             #f.write(f'{datetime.now()} {message.author}: {message.content} \n')
         if message.attachments:
@@ -83,22 +91,36 @@ class MyClient(discord.Client):
 
             open(filename, 'wb').write(img_data)            
 
-            #print ("the last message had attachments!")
+            print (bcolors.OKCYAN,"Processing attachments!",bcolors.ENDC)
 
             
             text = pytesseract.image_to_string(filename)
-            print (text)
-            os.remove(filename)
+            print (bcolors.OKGREEN, "OCR readings:", bcolors.ENDC, text)
+            
+            
             if words_in_string(word, text):
                 await message.delete()
+                print (bcolors.WARNING, "Keyword found in OCR reading! Deleting chat message.", bcolors.ENDC)
+                if store_flagged == 'False'
+                    os.remove(filename)
+            else: os.remove(filename)
             
                 
         
 
 
-
-
 client = MyClient(intents=intents)
 client.run(token)
 
+# COLORED TEXT    
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKCYAN = '\033[96m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
